# docs/swagger.yaml
basePath: /api
definitions:
  models.BaseID:
    properties:
      id:
        example: 479b5b5f-81b1-4669-91a5-b5bf69e597c6
        type: string
    type: object
  models.Note:
    properties:
      content:
        type: string
      createdAt:
        example: 2006-01-02T15:04:05Z07:00
        type: string
      deletedAt:
        example: 2006-01-02T15:04:05Z07:00
        format: date
        type: string
      id:
        example: 479b5b5f-81b1-4669-91a5-b5bf69e597c6
        type: string
      title:
        type: string
      updatedAt:
        example: 2006-01-02T15:04:05Z07:00
        type: string
      user:
        $ref: '#/definitions/models.User'
      userId:
        type: string
      visibility:
        $ref: '#/definitions/models.NoteVisibility'
    type: object
  models.NoteVisibility:
    enum:
    - public
    - private
    type: string
    x-enum-varnames:
    - NotePublic
    - NotePrivate
...
  models.Sort:
    enum:
    - asc
    - desc
    type: string
    x-enum-varnames:
    - ASC
    - DESC
  models.User:
    properties:
      createdAt:
        example: 2006-01-02T15:04:05Z07:00
        type: string
      deletedAt:
        example: 2006-01-02T15:04:05Z07:00
        format: date
        type: string
      email:
        example: chestaardi4@gmail.com
        type: string
      id:
        example: 479b5b5f-81b1-4669-91a5-b5bf69e597c6
        type: string
      status:
        $ref: '#/definitions/models.UserStatus'
      updatedAt:
        example: 2006-01-02T15:04:05Z07:00
        type: string
      username:
        example: chesta_ardiona
        type: string
    type: object
  models.UserStatus:
    enum:
    - active
    - pending
    type: string
    x-enum-varnames:
    - UserActive
    - UserPending
...
  payloads.RequestCreateNote:
    properties:
      content:
        type: string
      title:
        type: string
      visibility:
        $ref: '#/definitions/models.NoteVisibility'
    required:
    - content
    - title
    - visibility
    type: object
...
  payloads.RequestDeleteManyNote:
    properties:
      ids:
        items:
          type: string
        minItems: 1
        type: array
    required:
    - ids
    type: object
  payloads.RequestGetOneNote:
    properties:
      id:
        type: string
    required:
    - id
    type: object
  payloads.RequestRestoreManyNote:
    properties:
      ids:
        items:
          type: string
        minItems: 1
        type: array
      sort:
        allOf:
        - $ref: '#/definitions/models.Sort'
        description: default = desc
    required:
    - ids
    type: object
...
  reply.ErrorPayload:
    properties:
      code:
        description: Machine-readable error code
        type: string
      details:
        description: Optional debug details
        type: string
      fields:
        allOf:
        - $ref: '#/definitions/reply.FieldsError'
        description: Fields causing the error (if any)
      message:
        description: Human-readable message
        type: string
    type: object
  reply.FieldsError:
    additionalProperties:
      type: string
    type: object
  replylib.CodeError:
    enum:
    - NOT_FOUND
    - SERVER_ERROR
    - BAD_REQUEST
    - BAD_GATEWAY
    - UNAUTHORIZED
    - CONFLICT
    - FORBIDDEN
    - UNPROCESSABLE_ENTITY
    - TOO_MANY_REQUESTS
    - SERVICE_UNAVAILABLE
    - GATEWAY_TIMEOUT
    - METHOD_NOT_ALLOWED
    - NOT_ACCEPTABLE
    - REQUEST_TIMEOUT
    - PAYLOAD_TOO_LARGE
    - UNSUPPORTED_MEDIA_TYPE
    - GONE
    - NOT_IMPLEMENTED
    type: string
    x-enum-varnames:
    - SwagCodeNotFound
    - SwagCodeServerError
    - SwagCodeBadRequest
    - SwagCodeBadGateway
    - SwagCodeUnauthorized
    - SwagCodeConflict
    - SwagCodeForbidden
    - SwagCodeUnprocessableEntity
    - SwagCodeTooManyRequests
    - SwagCodeServiceUnavailable
    - SwagCodeGatewayTimeout
    - SwagCodeMethodNotAllowed
    - SwagCodeNotAcceptable
    - SwagCodeRequestTimeout
    - SwagCodePayloadTooLarge
    - SwagCodeUnsupportedMedia
    - SwagCodeGone
    - SwagCodeNotImplemented
  replylib.Envelope:
    properties:
      data: {}
      meta:
        $ref: '#/definitions/replylib.Meta'
    type: object
  replylib.Meta:
    properties:
      debug: {}
      information:
        type: string
      pagination:
        $ref: '#/definitions/replylib.Pagination'
      status:
        allOf:
        - $ref: '#/definitions/replylib.status'
        description: SUCCESS or ERROR
        example: SUCCESS
      timestamp:
        example: 2006-01-02T15:04:05Z07:00
        type: string
      tokens:
        additionalProperties:
          type: string
        type: object
    type: object
  replylib.Pagination:
    properties:
      current:
        description: current offset
        type: integer
      hasHext:
        description: true if data more than replied
        type: boolean
      next:
        description: if hasHext is false, next is 0
        type: integer
    type: object
  replylib.status:
    enum:
    - SUCCESS
    - ERROR
    type: string
    x-enum-varnames:
    - Success
    - Error
...
host: localhost:8080
info:
  contact: {}
  description: This is an API used for manages home server (ubuntu/debian).
  title: Homedy API
  version: "1.0"
paths:
...
  /notes:
    delete:
      parameters:
      - description: param of notes` identification
        in: body
        name: param
        required: true
        schema:
          $ref: '#/definitions/payloads.RequestDeleteManyNote'
      produces:
      - application/json
      responses:
        "200":
          description: data is null
          schema:
            $ref: '#/definitions/replylib.Envelope'
        default:
          description: ""
          schema:
            allOf:
            - $ref: '#/definitions/replylib.Envelope'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/reply.ErrorPayload'
                  - properties:
                      code:
                        $ref: '#/definitions/replylib.CodeError'
                    type: object
              type: object
      summary: Soft delete one client's existing notes
      tags:
      - note
    get:
      parameters:
      - in: query
        minimum: 0
        name: offset
        type: integer
      - in: query
        name: recycled
        type: boolean
      - description: default = desc
        enum:
        - asc
        - desc
        in: query
        name: sort
        type: string
        x-enum-varnames:
        - ASC
        - DESC
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/replylib.Envelope'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Note'
                  type: array
              type: object
        default:
          description: ""
          schema:
            allOf:
            - $ref: '#/definitions/replylib.Envelope'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/reply.ErrorPayload'
                  - properties:
                      code:
                        $ref: '#/definitions/replylib.CodeError'
                    type: object
              type: object
      summary: Get all client's existing notes with query
      tags:
      - note
    post:
      consumes:
      - application/json
      parameters:
      - description: new note data
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/payloads.RequestCreateNote'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/replylib.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/models.Note'
              type: object
        default:
          description: ""
          schema:
            allOf:
            - $ref: '#/definitions/replylib.Envelope'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/reply.ErrorPayload'
                  - properties:
                      code:
                        $ref: '#/definitions/replylib.CodeError'
                    type: object
              type: object
      summary: Create one new note
      tags:
      - note
  /notes/{id}:
    delete:
      parameters:
      - in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: deleted id in server
          schema:
            allOf:
            - $ref: '#/definitions/replylib.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/models.BaseID'
              type: object
        default:
          description: ""
          schema:
            allOf:
            - $ref: '#/definitions/replylib.Envelope'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/reply.ErrorPayload'
                  - properties:
                      code:
                        $ref: '#/definitions/replylib.CodeError'
                    type: object
              type: object
      summary: Soft delete one client's existing note
      tags:
      - note
    get:
      parameters:
      - in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/replylib.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/models.Note'
              type: object
        default:
          description: ""
          schema:
            allOf:
            - $ref: '#/definitions/replylib.Envelope'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/reply.ErrorPayload'
                  - properties:
                      code:
                        $ref: '#/definitions/replylib.CodeError'
                    type: object
              type: object
      summary: Get client's existing notes by id
      tags:
      - note
    put:
      consumes:
      - application/json
      parameters:
      - description: updated note
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/payloads.RequestGetOneNote'
      - example: 479b5b5f-81b1-4669-91a5-b5bf69e597c6
        in: path
        name: id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: updated note from server
          schema:
            allOf:
            - $ref: '#/definitions/replylib.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/models.Note'
              type: object
        default:
          description: ""
          schema:
            allOf:
            - $ref: '#/definitions/replylib.Envelope'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/reply.ErrorPayload'
                  - properties:
                      code:
                        $ref: '#/definitions/replylib.CodeError'
                    type: object
              type: object
      summary: Update client's existing note
      tags:
      - note
  /notes/restore:
    patch:
      parameters:
      - description: param of notes` identification
        in: body
        name: param
        required: true
        schema:
          $ref: '#/definitions/payloads.RequestRestoreManyNote'
      produces:
      - application/json
      responses:
        "200":
          description: restored notes data
          schema:
            allOf:
            - $ref: '#/definitions/replylib.Envelope'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Note'
                  type: array
              type: object
        default:
          description: ""
          schema:
            allOf:
            - $ref: '#/definitions/replylib.Envelope'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/reply.ErrorPayload'
                  - properties:
                      code:
                        $ref: '#/definitions/replylib.CodeError'
                    type: object
              type: object
      summary: Restore existing share by name
      tags:
      - note
  /notes/restore/{id}:
    patch:
      parameters:
      - in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: restored note data
          schema:
            allOf:
            - $ref: '#/definitions/replylib.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/models.Note'
              type: object
        default:
          description: ""
          schema:
            allOf:
            - $ref: '#/definitions/replylib.Envelope'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/reply.ErrorPayload'
                  - properties:
                      code:
                        $ref: '#/definitions/replylib.CodeError'
                    type: object
              type: object
      summary: Restore one client's existing note
      tags:
      - note
...
swagger: "2.0"